# docs/openapi.yaml
basePath: /api/v1
definitions:
...
  dto.ArticleAuthorDTO:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  dto.ArticleCategoryDTO:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
  dto.ArticleDTO:
    properties:
      author:
        $ref: '#/definitions/dto.ArticleAuthorDTO'
      category:
        $ref: '#/definitions/dto.ArticleCategoryDTO'
      category_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_user_generated:
        type: boolean
      moderation_status:
        type: string
      slug:
        type: string
      status:
        type: string
      thumbnail:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  dto.CreateUserArticleRequest:
    properties:
      category_id:
        type: integer
      content:
        type: string
      thumbnail:
        type: string
      title:
        type: string
    required:
    - category_id
    - content
    - title
    type: object
...
  dto.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      page:
        type: integer
      success:
        type: boolean
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.UpdateUserArticleRequest:
    properties:
      category_id:
        type: integer
      content:
        type: string
      thumbnail:
        type: string
      title:
        type: string
    required:
    - category_id
    - content
    - title
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ruangtenang.id
    name: API Support
  description: API untuk aplikasi Ruang Tenang - Platform Kesehatan Mental
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ruang Tenang API
  version: "1.0"
paths:
...
  /articles:
    get:
      description: Get paginated list of published articles with optional filtering
      parameters:
      - description: Filter by category ID
        in: query
        name: category_id
        type: integer
      - description: Search by title
        in: query
        name: search
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse'
      summary: Get articles list
      tags:
      - Articles
  /articles/{slug}:
    get:
      description: Get full article details by slug (only published)
      parameters:
      - description: Article Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ArticleDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Get article by slug
      tags:
      - Articles
...
  /my-articles:
    get:
      description: Get paginated list of articles owned by the authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search own article titles
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: Get user's own articles
      tags:
      - Articles
    post:
      consumes:
      - application/json
      description: Create a new article as the authenticated user
      parameters:
      - description: Article data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserArticleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Create a new article
      tags:
      - Articles
  /my-articles/{slug}:
    delete:
      description: Delete an article owned by the authenticated user
      parameters:
      - description: Article Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Delete user's own article
      tags:
      - Articles
    get:
      description: Get full article details (including own unpublished articles)
      parameters:
      - description: Article Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ArticleDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Get article by slug for authenticated user
      tags:
      - Articles
    put:
      consumes:
      - application/json
      description: Update an article owned by the authenticated user
      parameters:
      - description: Article Slug
        in: path
        name: slug
        required: true
        type: string
      - description: Article data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserArticleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Update user's own article
      tags:
      - Articles
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"